Fixed tests on windows
